# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GUM Internships API
  version: 1.0.0
  description: >
    MVP API contract. Every response uses the envelope {success, data, error, meta}.
    DB is snake_case; this API is camelCase. Auth = Bearer JWT (access 15m, refresh 30d).
    Pagination on list endpoints: ?page=&limit= (default 20, max 100).
    Endpoints marked "Public" need no token. x-required-roles lists allowed roles.
servers:
  - url: https://api.gum-internships.example.com/v1
    description: Production
  - url: https://staging-api.gum-internships.example.com/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /enrollments/{enrollmentId}/tasks:
    get:
      tags: [Enrollments]
      summary: Project tasks with my submission state per task
      x-required-roles: [student]
      parameters:
        - { name: enrollmentId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            task: { $ref: '#/components/schemas/ProjectTask' }
                            latestSubmission: { oneOf: [{ $ref: '#/components/schemas/Submission' }, { type: 'null' }] }
...
  /projects/{projectId}/tasks:
    post:
      tags: [Projects & Submissions]
      summary: Add weekly task (rubric JSON, allowed submission types)
      x-required-roles: [instructor, moderator, super_admin]
      parameters:
        - { name: projectId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weekNumber, title]
              properties:
                weekNumber: { type: integer, minimum: 1 }
                title: { type: string }
                instructions: { type: string }
                allowedSubmissionTypes: { type: array, items: { type: string, enum: [file, github_url, live_url] } }
                maxScore: { type: number, default: 100 }
                rubric: { type: array, items: { type: object, properties: { criterion: { type: string }, weight: { type: number }, maxPoints: { type: number } } } }
                dueOffsetDays: { type: integer, minimum: 0 }
                isMandatory: { type: boolean, default: true }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/ProjectTask' } }
...
  /tasks/{taskId}/submissions:
    post:
      tags: [Projects & Submissions]
      summary: Submit work (new version; 409 if a version is pending review)
      x-required-roles: [student]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollmentId, submissionType]
              properties:
                enrollmentId: { type: integer, format: int64 }
                submissionType: { type: string, enum: [file, github_url, live_url] }
                fileUrl: { type: string, description: Bunny path from presigned upload }
                urlValue: { type: string }
                notes: { type: string }
      responses:
        '201':
          description: Submitted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Submission' } }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /instructor/review-queue:
    get:
      tags: [Projects & Submissions]
      summary: Pending submissions across own internships (oldest first)
      x-required-roles: [instructor, moderator, super_admin]
      parameters:
        - { name: internshipId, in: query, schema: { type: integer, format: int64 } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Submission' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
  /submissions/{submissionId}/review:
    post:
      tags: [Projects & Submissions]
      summary: Review a submission (rubric scores; feedback required unless approved)
      x-required-roles: [instructor, moderator, super_admin]
      parameters:
        - { name: submissionId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision, rubricScores]
              properties:
                decision: { type: string, enum: [approved, changes_requested, rejected] }
                rubricScores: { type: array, items: { type: object, required: [criterion, points], properties: { criterion: { type: string }, points: { type: number } } } }
                feedback: { type: string }
      responses:
        '201':
          description: Reviewed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/SubmissionReview' } }
        '409': { $ref: '#/components/responses/Conflict' }

  # ============================================================ QUIZZES
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
...
    Conflict:
      description: 'State conflict, e.g. ALREADY_ENROLLED, BATCH_FULL, SUBMISSION_PENDING_REVIEW'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    Envelope:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, const: true }
        data: {}
        error: { type: 'null' }
        meta: { type: object }
    ErrorEnvelope:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, const: false }
        data: { type: 'null' }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: 'Stable machine code, e.g. VALIDATION_ERROR, FORBIDDEN, BATCH_FULL' }
            message: { type: string }
            details: { description: Optional zod issue list }
    PaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    ProjectTask:
      type: object
      properties:
        id: { type: integer, format: int64 }
        projectId: { type: integer, format: int64 }
        weekNumber: { type: integer }
        title: { type: string }
        instructions: { type: [string, 'null'] }
        allowedSubmissionTypes: { type: array, items: { type: string, enum: [file, github_url, live_url] } }
        maxScore: { type: number }
        rubric: { type: array, items: { type: object, properties: { criterion: { type: string }, weight: { type: number }, maxPoints: { type: number } } } }
        dueDate: { type: [string, 'null'], format: date, description: Resolved from batch start + dueOffsetDays }
        isMandatory: { type: boolean }
    Submission:
      type: object
      properties:
        id: { type: integer, format: int64 }
        taskId: { type: integer, format: int64 }
        enrollmentId: { type: integer, format: int64 }
        version: { type: integer }
        submissionType: { type: string, enum: [file, github_url, live_url] }
        fileUrl: { type: [string, 'null'], description: Signed URL when served }
        urlValue: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        status: { type: string, enum: [submitted, under_review, approved, changes_requested, rejected] }
        submittedAt: { type: string, format: date-time }
        review: { oneOf: [{ $ref: '#/components/schemas/SubmissionReview' }, { type: 'null' }] }
    SubmissionReview:
      type: object
      properties:
        id: { type: integer, format: int64 }
        submissionId: { type: integer, format: int64 }
        reviewerName: { type: string }
        rubricScores: { type: array, items: { type: object, properties: { criterion: { type: string }, points: { type: number }, maxPoints: { type: number } } } }
        totalScore: { type: [number, 'null'] }
        decision: { type: string, enum: [approved, changes_requested, rejected] }
        feedback: { type: [string, 'null'] }
        reviewedAt: { type: string, format: date-time }
...